% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bgm_df}
\alias{bgm_df}
\title{bgm_df}
\format{A data frame with 324 rows and 5 variables:
\describe{
  \item{\code{lat}}{double. Latitude}
  \item{\code{long}}{double. Longitude}
  \item{\code{inside_lat}}{double. Latitude of ("mid"-) point inside the polygon.}
  \item{\code{inside_long}}{double. Longitude of ("mid"-) point inside the polygon.}
  \item{\code{polygon}}{double. Polygonid.}
}}
\usage{
bgm_df
}
\description{
bgm file converted to tidy dataframe.
}
\keyword{datasets}

