% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-to-layer.R
\name{depth_to_layer}
\alias{depth_to_layer}
\title{Convert mean_depth from ECOHAM output to ATLANTIS layer id.}
\usage{
depth_to_layer(polygon, mean_depth, nominal_dz)
}
\arguments{
\item{polygon}{int Vector of ATLANTIS polygon id.}

\item{mean_depth}{num Vector of ECOHAM mean depth.}

\item{nominal_dz}{dataframe Tidy dataframe with columns polygon, layer, max_nominal_dz.
See \link{create-example-dfs.R} for further details.}
}
\value{
Dataframe.
}
\description{
Convert mean_depth from ECOHAM output to ATLANTIS layer id.
}
\examples{
polygon <- c(12, 15)
mean_depth <- c(5, 12.5)
nominal_dz <- ecocat::nominal_dz_df
}

