% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoham_layout}
\alias{ecoham_layout}
\title{ecoham_layout}
\format{A data frame with 7216 rows and 3 variables:
\describe{
  \item{\code{longitude}}{double. longitude.}
  \item{\code{latitude}}{double. latitude.}
  \item{\code{ecoham_id}}{integer. ECOHAM box id.}
  \item{\code{area}}{double. grid area in square meters.}
}}
\usage{
ecoham_layout
}
\description{
mid points in latitude and longitude for each ecoham box.
}
\keyword{datasets}
