% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoham_vol}
\alias{ecoham_vol}
\title{Reference dataframe for volume per ecoham grid.}
\format{A data frame with 8819 rows and 3 variables:
\describe{
  \item{\code{depth}}{double. depth:units = "meters".}
  \item{\code{ecoham_out}}{double. vol:units = "m3".}
  \item{\code{ecoham_id}}{integer. grid cell.}
}}
\usage{
ecoham_vol
}
\description{
Volume in m^3 for each ECOHAM grid and depth layer.
}
\keyword{datasets}

