% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-unit.R
\name{extract_unit}
\alias{extract_unit}
\title{Extract units for each variable in a netcdf file}
\usage{
extract_unit(nc, attribute = "units")
}
\arguments{
\item{nc}{connection to the netcdf file to read in.}

\item{attribute}{either a character string of the name of the attribute (deault = "units") or the
attribute id (starting with 0).}
}
\description{
Extract units for each variable in a netcdf file
}
\examples{
nc <- system.file(package = "ecocat", "extdata/volume.nc")
extract_unit(nc)
}
