% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect-ecocat.R
\name{intersect_ecocat}
\alias{intersect_ecocat}
\title{Calculate overlap between ecoham boxes and atlantis polygons.}
\usage{
intersect_ecocat(bgm_df, ecoham_layout)
}
\arguments{
\item{bgm_df}{Dataframe generated with atlantistools::convert_bgm().}

\item{ecoham_layout}{Dataframe with columns longitude, latidude (box mid points) and the ecoham_id for
each box. See \link{create-example-dfs.R} for details.}
}
\description{
Calculate overlap between ecoham boxes and atlantis polygons.
}
\examples{
intersect_ecocat(bgm_df, ecoham_layout)
}

