% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nicemap_df}
\alias{nicemap_df}
\title{nicemap_df}
\format{A data frame with 7216 rows and 3 variables:
\describe{
  \item{\code{id_x}}{integer. Horizontal position of the potential ECOHAM box.}
  \item{\code{id_y}}{integer. Vertical position of the potential ECOHAM box.}
  \item{\code{poly_code}}{character. Value of the box code.}
  \item{\code{ecoham_id}}{integer. ECOHAM box id.}
  \item{\code{polygon}}{double. ATLANTIS polygon id.}
}}
\usage{
nicemap_df
}
\description{
nicemap converted to tidy dataframe.
}
\keyword{datasets}
