% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nominal_dz_df}
\alias{nominal_dz_df}
\title{Cumulative sum of nominal_dz from GNS-Atlantis model.}
\format{A data frame with 95 rows and 3 variables:
\describe{
  \item{\code{polygon}}{integer. ATLANTIS polygon.}
  \item{\code{layer}}{double. ATLANTIS layer.}
  \item{\code{max_nominal_dz}}{double. maximum cimulative nominal_dz per polygon and layer.}
}}
\usage{
nominal_dz_df
}
\description{
Cumulative nominal_dz.
}
\keyword{datasets}

