% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_depth_df}
\alias{poly_depth_df}
\title{Unique combination between ATLANTIS polygons and ECOHAM mean depths.}
\format{A data frame with 40 rows and 2 variables:
\describe{
  \item{\code{depth}}{double. mean depth of ECOHAM grid cells.}
  \item{\code{polygon}}{double. ATLANTIS polygon ids.}
}}
\usage{
poly_depth_df
}
\description{
Polygons and depths.
}
\keyword{datasets}
