% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth-to-layer.R
\name{depth_to_layer}
\alias{depth_to_layer}
\title{Convert mean_depth from ECOHAM output to ATLANTIS layer id.}
\usage{
depth_to_layer(poly_depth, nominal_dz)
}
\arguments{
\item{poly_depth}{dataframe Unique combinations between ATLANTIS polygons and ECOHAM mean depths.
Layer assignment does not work in case the true mininum mean depth is not present here.}

\item{nominal_dz}{dataframe Tidy dataframe with columns polygon, layer, max_nominal_dz.
See \link{nominal_dz_df} for further details.}
}
\value{
Dataframe.
}
\description{
Convert mean_depth from ECOHAM output to ATLANTIS layer id.
}
\examples{
nc <- system.file(package = "ecocat", "extdata/d1.nc")
df <- ecocat(nc)
df <- dplyr::select(df, depth, polygon)
df <- unique(df)
poly_depth <- df
nominal_dz <- ecocat::nominal_dz_df
}

