% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecocat.R
\name{ecocat}
\alias{ecocat}
\title{Convert ECOHAM output to ATLANTIS output.}
\usage{
ecocat(nc, nicemap = nicemap_df)
}
\arguments{
\item{nc}{connection to the netcdf file to read in.}

\item{nicemap}{dataframe of the nicemap.}
}
\value{
Dataframe.
}
\description{
Assign spatial output from ECOHAM simulation to ATLANTIS polygons.
}
\examples{
nc <- system.file(package = "ecocat", "extdata/volume.nc")
df <- ecocat(nc)

nc <- system.file(package = "ecocat", "extdata/d1.nc")
df <- ecocat(nc)
}

