% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecocat.R
\name{ecocat}
\alias{ecocat}
\title{Convert ECOHAM output to ATLANTIS output.}
\usage{
ecocat(eco_out, col, nicemap_df)
}
\arguments{
\item{eco_out}{dataframe generated with \link{eco-to-tidy}.}

\item{col}{column to convert from ecoham output to atlantis output.}

\item{nicemap_df}{dataframe of the nicemap.}
}
\value{
Dataframe.
}
\description{
Assign spatial output from ECOHAM simulation to ATLANTIS polygons.
}
\examples{
eco_out <- ref_vol
ecocat(ref_vol, col = "vol", nicemap_df)
}

