% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecocat.R
\name{ecocat}
\alias{ecocat}
\title{Convert ECOHAM output to ATLANTIS output.}
\usage{
ecocat(nc, nicemap = ecocat::nicemap_df, nominal_dz = ecocat::nominal_dz_df)
}
\arguments{
\item{nc}{connection to the netcdf file to read in.}

\item{nicemap}{dataframe of the nicemap.}

\item{nominal_dz}{dataframe cumulative water column thickness per ATLANTIS polygon and layer.}
}
\value{
Dataframe.
}
\description{
Assign spatial output from ECOHAM simulation to ATLANTIS polygons.
}
\examples{
nc <- system.file(package = "ecocat", "extdata/volume.nc")
df <- ecocat(nc)

nc <- system.file(package = "ecocat", "extdata/d1.nc")
df <- ecocat(nc)
}

