% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect-ecocat.R
\name{intersect_ecocat}
\alias{intersect_ecocat}
\title{Calculate overlap between ecoham boxes and atlantis polygons.}
\usage{
intersect_ecocat(atlantis_bgm, eco_layout = ecocat::ecoham_layout)
}
\arguments{
\item{atlantis_bgm}{bgm Geometry file of the ATLANTIS model.}

\item{eco_layout}{dataframe ECOHAM grid layout. See  \code{\link{ecoham_layout}} for further details.}
}
\value{
Dataframe with columns "id", "ecoham_id", "polygon", "area.x", "longitude", "latitude",
"area.y" and "wf_area".
}
\description{
Calculate overlap between ecoham boxes and atlantis polygons.
}
\examples{
\dontrun{
atlantis_bgm <- system.file(package = "ecocat", "extdata/NorthSea.bgm")
overlap <- intersect_ecocat(atlantis_bgm, ecoham_layout)
}
}
