% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_vol}
\alias{ref_vol}
\title{Reference dataframe for volume per ecoham grid.}
\format{A data frame with 17638 rows and 6 variables:
\describe{
  \item{\code{longitude}}{double. ongitude:units = "degrees_east".}
  \item{\code{latitude}}{double. latitude:units = "degrees_north".}
  \item{\code{depth}}{double. depth:units = "meters".}
  \item{\code{time}}{double. time:units = "days since 2000-1-1 00:00:00".}
  \item{\code{vol}}{double. vol:units = "m3".}
  \item{\code{ecoham_id}}{integer. grid cell.}
}}
\usage{
ref_vol
}
\description{
Volume in m³ for each ECOHAM grid and depth layer.
}
\keyword{datasets}

